+++ Demo/Demo/Demo/DemoApp.swift
@@ -17,7 +17,7 @@ struct DemoApp: App {
     
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            HomeScreen()
         }
     }
 }